requirements.txt
asgiref==3.7.2
astroid==2.11.7
autopep8==1.7.0
backports.entry-points-selectable==1.1.1
botocore==1.31.55
dill==0.3.5.1
distlib==0.3.3
Django==4.2.5
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
environ==1.0
filelock==3.4.0
git-remote-codecommit==1.17
inflection==0.5.1
isort==5.10.1
jmespath==1.0.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
packaging==23.2
platformdirs==2.4.0
psycopg2==2.9.9
pycodestyle==2.9.1
pygraphviz==1.11
PyJWT==2.8.0
pylint==2.14.5
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.16
whitenoise==6.6.0
wrapt==1.14.1
